Remove ScriptIterator, can easily replace ScriptParser. Fix tests
