+++ tests/Script/ScriptParserTest.php
@@ -68,11 +68,16 @@ class ScriptParserTest extends \PHPUnit_Framework_TestCase
     public function testPush($script, $expectedOp, $expectedPushData, $result)
     {
         $parser = ScriptFactory::fromHex($script)->getScriptParser();
-        $pushdata = new Buffer();
-        $this->assertEquals($result, $parser->next($opCode, $pushdata));
-        $this->assertSame($expectedOp, $opCode);
-        if ($pushdata->getSize() > 0) {
-            $this->assertSame($expectedPushData, $pushdata->getBinary());
+
+        $result = $parser->next();
+
+        if ($result !== null) {
+            if ($result->isPush()) {
+                $data = $result->getData();
+                if ($data->getSize() > 0) {
+                    $this->assertSame($expectedPushData, $data->getBinary());
+                }
+            }
         }
     }
